Move the model tests to a water schema folder
