+++ test/models/water/region.model.test.js
@@ -8,15 +8,15 @@ const { describe, it, beforeEach } = exports.lab = Lab.script()
 const { expect } = Code
 
 // Test helpers
-const BillingBatchHelper = require('../support/helpers/billing-batch.helper.js')
-const BillingBatchModel = require('../../app/models/billing-batch.model.js')
-const DatabaseHelper = require('../support/helpers/database.helper.js')
-const LicenceHelper = require('../support/helpers/licence.helper.js')
-const LicenceModel = require('../../app/models/licence.model.js')
-const RegionHelper = require('../support/helpers/region.helper.js')
+const BillingBatchHelper = require('../../support/helpers/water/billing-batch.helper.js')
+const BillingBatchModel = require('../../../app/models/water/billing-batch.model.js')
+const DatabaseHelper = require('../../support/helpers/database.helper.js')
+const LicenceHelper = require('../../support/helpers/water/licence.helper.js')
+const LicenceModel = require('../../../app/models/water/licence.model.js')
+const RegionHelper = require('../../support/helpers/water/region.helper.js')
 
 // Thing under test
-const RegionModel = require('../../app/models/region.model.js')
+const RegionModel = require('../../../app/models/water/region.model.js')
 
 describe('Region model', () => {
   let testRecord
